+++ src/gudrun_classes/sample.py
@@ -108,7 +108,7 @@ class Sample:
         self.periodNumber = 0
         self.dataFiles = DataFiles([], "SAMPLE")
         self.forceCalculationOfCorrections = False
-        self.composition = Composition([], "SAMPLE")
+        self.composition = Composition("SAMPLE")
         self.geometry = Geometry.SameAsBeam
         self.upstreamThickness = 0.0
         self.downstreamThickness = 0.0
